refactor: moved sort functions to utils file
